mc status bot/requirements.txt
discord.py>=1.5.0
mcstatus
pyyaml
